@@ -1,11 +1,11 @@
 from datetime import datetime
 from enum import Enum
-from typing import List, Optional
+from typing import Optional
 
 from jkit.msgspec_constraints import PositiveInt, UserName, UserSlug, UserUploadedUrl
-from sshared.mongo import MODEL_META, Document, Index
+from sshared.mongo import Document, Index
 
-from utils.db import JIANSHU_DB
+from utils.mongo import JIANSHU_DB
 
 
 class JianshuUserStatus(Enum):
@@ -13,13 +13,13 @@ class JianshuUserStatus(Enum):
     INACCESSABLE = "INACCESSIBLE"
 
 
-class UserDocument(Document, **MODEL_META):
+class UserDocument(Document, frozen=True):
     slug: UserSlug
     status: JianshuUserStatus
     updated_at: datetime
     id: Optional[PositiveInt]
     name: Optional[UserName]
-    history_names: List[UserName]
+    history_names: list[UserName]
     avatar_url: Optional[UserUploadedUrl]
 
     class Meta:  # type: ignore
