@@ -6,19 +6,19 @@ from jkit.msgspec_constraints import (
     PositiveFloat,
     PositiveInt,
 )
-from sshared.mongo import MODEL_META, Document, Field, Index
+from sshared.mongo import Document, Field, Index
 
-from utils.db import JPEP_DB
+from utils.mongo import JPEP_DB
 
 
-class AmountField(Field, **MODEL_META):
+class AmountField(Field, frozen=True):
     total: PositiveInt
     traded: NonNegativeInt
     tradable: NonNegativeInt
     minimum_trade: PositiveInt
 
 
-class FTNTradeOrderDocument(Document, **MODEL_META):
+class FTNTradeOrderDocument(Document, frozen=True):
     fetch_time: datetime
     id: PositiveInt
     published_at: datetime
